Fix login logic to use api module
