+++ .eslintrc.json
@@ -107,6 +107,7 @@
     "react/jsx-props-no-spreading": "off",
     "react/require-default-props": "off",
     "no-use-before-define": "off",
-    "@typescript-eslint/no-use-before-define": ["error"]
+    "@typescript-eslint/no-use-before-define": ["error"],
+    "@typescript-eslint/return-await": "off"
   }
 }
