+++ app/services/management/auth.js
@@ -2,6 +2,8 @@ import Keycloak from 'keycloak-js';
 import * as jwt from 'jsonwebtoken';
 import { $http } from 'ngimport';
 
+import api from '@/services/api';
+
 const externalGroups = ['clinician', 'collaborator', 'external analyst'];
 
 const keycloak = Keycloak({
@@ -60,14 +62,9 @@ const isAuthorized = (authorizationToken) => {
 /**
  * Gets the user object from the api
  */
-const getUser = async (token) => {
+const getUser = async () => {
   try {
-    // Token passed on login
-    if (token) {
-      $http.defaults.headers.common.Authorization = token;
-    }
-    const resp = await $http.get(`${window._env_.API_BASE_URL}/user/me`);
-    return resp.data;
+    return api.get('/user/me').request();
   } catch (err) {
     return null;
   }
