+++ app/views/ReportsView/index.jsx
@@ -40,7 +40,7 @@ function ReportsTableComponent(props) {
       states = 'reviewed,archived';
     }
 
-    const { reports } = await api.get(`/reports?states=${states}`, {});
+    const { reports } = await api.get(`/reports?states=${states}`, {}).request();
 
     setRowData(reports.map((report) => {
       const [analyst] = report.users
